+++ src/main/scala/edu/ie3/simona/model/participant/HpModel.scala
@@ -107,20 +107,46 @@ class HpModel private (
       state: HpState
   ): FlexOptions = {
     val wasRunningLastOp = state.lastHpOperatingPoint.activePower > zeroKW
-
     // Determining the operation point and limitations at this tick
-    val (turnOn, canOperate, canBeOutOfOperation) =
-      determineHpOperatingOptions(
-        state.thermalGridState,
-        state.thermalDemands,
-        wasRunningLastOp,
-      )
-
-    MinMaxFlexOptions(
-      if (turnOn) sRated.toActivePower(cosPhiRated) else zeroKW,
-      if (canBeOutOfOperation) zeroKW else sRated.toActivePower(cosPhiRated),
-      if (canOperate) sRated.toActivePower(cosPhiRated) else zeroKW,
+    val (turnOn, canOperate, canBeOutOfOperation) = determineHpOperatingOptions(
+      state.thermalGridState,
+      state.thermalDemands,
+      wasRunningLastOp,
     )
+
+    val (refPower, minPower) = (turnOn, canBeOutOfOperation) match {
+      case (true, true) =>
+        if (
+          state.lastHpOperatingPoint.activePower > zeroKW &&
+          state.thermalDemands.houseDemand.hasPossibleDemand &&
+          state.thermalGridState.storageState
+            .map(_.storedEnergy)
+            .getOrElse(zeroKWh) == zeroKWh
+        )
+          // if Hp was running last state AND there is demand from the house AND the storage is empty,
+          // we would like to keep that behaviour even in strict interpretation of flexibility we could
+          // be out of operation for flex reasons. Thus, we force Hp to run.
+          (sRated.toActivePower(cosPhiRated), sRated.toActivePower(cosPhiRated))
+        else {
+          (sRated.toActivePower(cosPhiRated), zeroKW)
+        }
+      case (true, false) =>
+        (sRated.toActivePower(cosPhiRated), sRated.toActivePower(cosPhiRated))
+      case (false, true) =>
+        (
+          zeroKW,
+          zeroKW,
+        )
+      case (false, false) =>
+        (
+          sRated.toActivePower(cosPhiRated),
+          sRated.toActivePower(cosPhiRated),
+        ) // should not be possible to reach
+    }
+
+    val maxPower = if (canOperate) sRated.toActivePower(cosPhiRated) else zeroKW
+
+    MinMaxFlexOptions(refPower, minPower, maxPower)
   }
 
   /** Calculate the active power behaviour of the model.